requirements.txt
Django
psycopg[binary]
djangorestframework
django-cors-headers

